+++ queue/src/Main.java
@@ -2,8 +2,8 @@ public class Main {
     public static void main(String[] args) {
         Queue queue = new Queue();
 
-        Client lucas = new Client("lucas", 22, 10000.0);
-        Client david = new Client("david", 16, 10000.0);
+        Client lucas = new Client("lucas", (short) 22, 10000.0);
+        Client david = new Client("david", (short) 16, 10000.0);
 
         queue.addClient(lucas);
         queue.addClient(david);
