+++ stack/src/Person.java
@@ -2,7 +2,7 @@ import java.util.Objects;
 
 public abstract class Person {
     private String name;
-    private Integer age;
+    private Short age;
     private Character genre;
 
     private Person next;
@@ -15,11 +15,11 @@ public abstract class Person {
         this.name = name;
     }
 
-    public Integer getAge() {
+    public Short getAge() {
         return age;
     }
 
-    public void setAge(Integer age) {
+    public void setAge(Short age) {
         this.age = age;
     }
